+++ rubato/utils/__init__.py
@@ -7,7 +7,6 @@ from .rb_math import Math
 from .noise import Noise
 from .vector import Vector
 from .display import Display
-from .surf import Surf
 from .color import Color
 from .font import Font
 from .rb_input import Input